+++ src/render/descriptor_pool.cc
@@ -45,7 +45,7 @@ void render::DescriptorPool::AllocateSet(VkDescriptorSetLayout descriptor_set_la
 
 void render::DescriptorPool::FreeSet(std::vector<VkDescriptorSet>& allocated_sets)
 {
-	vkFreeDescriptorSets(global_.logical_device, handle_, allocated_sets.size(), allocated_sets.data());
+	vkFreeDescriptorSets(global_.logical_device, handle_, u32(allocated_sets.size()), allocated_sets.data());
 }
 
 render::DescriptorPool::~DescriptorPool()
